third_party/WebKit/LayoutTests/bluetooth/idl-BluetoothDevice.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  assert_throws(null, function() { new BluetoothDevice(); },
                'the constructor should not be callable with "new"');
  assert_throws(null, function() { BluetoothDevice(); },
                'the constructor should not be callable');
}, 'BluetoothDevice IDL test');
</script>
